avro write benchmark
